WIP notebook that currently is not working because of changes in wetterdienst
The problem is that I had to update to 0.82 or 0.83 because of changes
in the files that are downloaded by wetterdienst from the DWD server.
These new versions conflict with installing pandas. My notebook still
relies on pandas because I wrote it with a fairly old version of
wetterdienst before they switched to polars. Hence, I cannot export from
polars to pandas and thus will need to change all pandas operations to
polar operations, which is a bit of a effort...